- don't cast complex expressions

+++ nsd/nsmain.c
@@ -265,7 +265,7 @@ Ns_Main(int argc, char *const*argv, Ns_ServerInitProc *initProc)
 
     if (mode == 'c') {
 	int i;
-        cmd.argv = ns_calloc((size_t)(argc - optind + 2), sizeof(char *));
+        cmd.argv = ns_calloc((size_t)argc - (size_t)optind + 2u, sizeof(char *));
         cmd.argc = 0;
         cmd.argv[cmd.argc++] = argv[0];
         for (i = optind; i < argc; i++) {
